chore(directory): 관심사에 맞게 api 관련 파일 이동
