+++ src/types/api.ts
@@ -1,0 +1,41 @@
+interface Movie {
+  adult: boolean;
+  backdrop_path: string;
+  genre_ids: number[];
+  id: number;
+  original_language: string;
+  original_title: string;
+  overview: string;
+  popularity: number;
+  poster_path: string;
+  release_date: string;
+  title: string;
+  video: boolean;
+  vote_average: number;
+  vote_count: number;
+}
+
+interface MovieResponse {
+  page: number;
+  results: Movie[];
+  total_pages: number;
+  total_results: number;
+}
+
+interface Genre {
+  id: number;
+  name: string;
+}
+
+interface MovieDetail {
+  id: number;
+  title: string;
+  poster_path: string;
+  backdrop_path: string;
+  release_date: string;
+  genres: Genre[];
+  vote_average: number;
+  overview: string;
+}
+
+export type { Movie, MovieResponse, Genre, MovieDetail };
